revert the code sharing approach, use simple repository structure to allow `charmcraft pack` in repository root
