+++ .github/workflows/ci.yaml
@@ -21,31 +21,10 @@ on:
 
 jobs:
   lint:
-    strategy:
-      matrix:
-        path:
-          - valkey-operator/kubernetes
-          - common
     name: Lint | ${{ matrix.path }}
-    runs-on: ubuntu-latest
-    timeout-minutes: 5
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v5
-      - name: Install tox & poetry
-        run: |
-          pipx install tox
-          pipx install poetry
-      - name: Run linters
-        working-directory: ${{ matrix.path }}
-        run: |
-          tox run -e lint
+    uses: canonical/data-platform-workflows/.github/workflows/lint.yaml@v40.0.2
 
   unit-test:
-    strategy:
-      matrix:
-        path:
-          - valkey-operator/kubernetes
     name: Unit test | ${{ matrix.path }}
     runs-on: ubuntu-latest
     timeout-minutes: 10
@@ -62,23 +41,39 @@ jobs:
       - name: Upload Coverage to Codecov
         uses: codecov/codecov-action@v4
 
+  lib-check:
+    name: Check libraries
+    runs-on: ubuntu-latest
+    timeout-minutes: 5
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+      - run: |
+          # Workaround for https://github.com/canonical/charmcraft/issues/1389#issuecomment-1880921728
+          touch requirements.txt
+      - name: Check libs
+        uses: canonical/charming-actions/check-libraries@2.7.0
+        continue-on-error: true
+        with:
+          credentials: ${{ secrets.CHARMHUB_TOKEN }}
+          github-token: ${{ secrets.GITHUB_TOKEN }}
+          use-labels: false
+          fail-build: ${{ github.event_name == 'pull_request' }}
+
   build:
     strategy:
       fail-fast: true
       matrix:
         path:
-          - valkey-operator/kubernetes
+          - .
     name: Build charm | ${{ matrix.path }}
     uses: canonical/data-platform-workflows/.github/workflows/build_charm.yaml@v40.0.2
     with:
       path-to-charm-directory: ${{ matrix.path }}
 
   integration-test:
-    strategy:
-      fail-fast: false
-      matrix:
-        path:
-          - valkey-operator/kubernetes
     name: Integration test | ${{ matrix.path }}
     needs:
       - lint
@@ -87,5 +82,4 @@ jobs:
     uses: ./.github/workflows/integration_test.yaml
     with:
       artifact-prefix: ${{ needs.build.outputs.artifact-prefix }}
-      path-to-charm-directory: ${{ matrix.path }}
     secrets: inherit
