--- common/common/__init__.py
@@ -1,5 +1,0 @@
-# Copyright 2025 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-"""Code shared between Valkey operator and Valkey cluster operator
-as well as between Kubernetes & machine charms"""
