--- common/common/lib_fetch_charmcraft.yaml
@@ -1,11 +1,0 @@
-# Copyright 2025 Canonical Ltd.
-# See LICENSE file for licensing details.
-#
-# This file is here only to ensure libraries can be fetched and updated.
-name: valkey-operator
-summary: Placeholder for lib refresh.
-description: Placeholder for lib refresh.
-
-type: charm
-platforms:
-  ubuntu@24.04:amd64:
