--- common/common/literals.py
@@ -1,8 +1,0 @@
-#!/usr/bin/env python3
-# Copyright 2025 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-"""Collection of global literals for the Valkey charms."""
-
-PEER_RELATION = "valkey-peers"
-STATUS_PEERS_RELATION = "status-peers"
