--- common/pyproject.toml
@@ -1,56 +1,0 @@
-[project]
-name = "common"
-version = "0.1.0"
-requires-python = ">=3.12" # switch to 3.14 once charm base is 26.04
-dependencies = [
-    "ops>=3.5.0",
-]
-exclude = [
-    "tests"
-]
-
-[build-system]
-requires = ["poetry-core>=2.0.0,<3.0.0"]
-build-backend = "poetry.core.masonry.api"
-
-[tool.poetry.group.charm-libs.dependencies]
-# data_platform_libs/v0/data_interfaces.py
-ops = ">=2.0.0"
-
-[tool.poetry.group.format]
-optional = true
-
-[tool.poetry.group.format.dependencies]
-ruff = "*"
-
-[tool.poetry.group.lint]
-optional = true
-
-[tool.poetry.group.lint.dependencies]
-ruff = "*"
-codespell = "*"
-shellcheck-py = "*"
-
-# Linting tools configuration
-[tool.ruff]
-line-length = 99
-lint.select = ["E", "W", "F", "C", "N", "D", "I001"]
-lint.extend-ignore = [
-    "D203",
-    "D204",
-    "D213",
-    "D215",
-    "D400",
-    "D404",
-    "D406",
-    "D407",
-    "D408",
-    "D409",
-    "D413",
-]
-lint.ignore = ["E501", "D107"]
-extend-exclude = ["__pycache__", "*.egg_info"]
-lint.per-file-ignores = { "tests/*" = ["D100", "D101", "D102", "D103", "D104"] }
-
-[tool.ruff.lint.mccabe]
-max-complexity = 10
