--- common/tox.ini
@@ -1,48 +1,0 @@
-# Copyright 2025 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-[tox]
-no_package = True
-skip_missing_interpreters = True
-env_list = format, lint, static, unit
-min_version = 4.0.0
-
-[vars]
-core_path = {tox_root}/common/core
-events_path = {tox_root}/common/events
-managers_path = {tox_root}/common/managers
-tests_path = {tox_root}/tests
-;lib_path = {tox_root}/lib/charms/operator_name_with_underscores
-all_path = {[vars]core_path} {[vars]events_path} {[vars]managers_path}
-
-[testenv]
-set_env =
-    PYTHONPATH = {[vars]core_path}:{[vars]events_path}:{[vars]managers_path}:{tox_root}/lib
-    PY_COLORS = 1
-allowlist_externals =
-    poetry
-
-[testenv:format]
-description = Apply coding style standards to code
-commands_pre =
-    poetry install --only format
-commands =
-    poetry lock
-    poetry run ruff check --fix {[vars]all_path}
-    poetry run ruff format {[vars]all_path}
-
-[testenv:lint]
-description = Check code against coding style standards
-allowlist_externals =
-    {[testenv]allowlist_externals}
-    find
-commands_pre =
-    poetry install --only lint
-commands =
-    poetry check --lock
-    # uncomment the following line if this charm owns a lib
-    # codespell {[vars]lib_path}
-    poetry run codespell {[vars]all_path} --skip "valkey.conf"
-    poetry run ruff check {[vars]all_path}
-    poetry run ruff format --check --diff {[vars]all_path}
-    find {[vars]all_path} -type f \( -name "*.sh" -o -name "*.bash" \) -exec poetry run shellcheck --color=always \{\} +
