+++ src/literals.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python3
+# Copyright 2025 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+"""Collection of global literals for the Valkey K8s charm."""
+
+CHARM = "valkey-k8s"
+CHARM_USER = "valkey"
+CONTAINER = "valkey"
+
+CONFIG_FILE = "/var/lib/valkey/valkey.conf"
+
+PEER_RELATION = "valkey-peers"
+STATUS_PEERS_RELATION = "status-peers"
