+++ tests/spread/test_charm.py/task.yaml
@@ -1,0 +1,9 @@
+summary: test_charm.py
+environment:
+  TEST_MODULE: test_charm.py
+systems:
+  - self-hosted-linux-amd64-noble-medium
+execute: |
+  tox run -e integration -- "tests/integration/k8s/$TEST_MODULE" --alluredir="$SPREAD_TASK/allure-results"
+artifacts:
+  - allure-results
