+++ tests/unit/conftest.py
@@ -1,0 +1,27 @@
+#!/usr/bin/env python3
+# Copyright 2026 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+import pytest
+from ops import testing
+
+
+@pytest.fixture(autouse=True)
+def mock_write_config_file(mocker):
+    mocker.patch("workload.ValkeyK8sWorkload.write_config_file")
+
+
+@pytest.fixture(autouse=True)
+def cloud_spec():
+    return testing.CloudSpec(
+        type="lxd",
+        endpoint="https://127.0.0.1:8443",
+        credential=testing.CloudCredential(
+            auth_type="clientcertificate",
+            attributes={
+                "client-cert": "foo",
+                "client-key": "bar",
+                "server-cert": "baz",
+            },
+        ),
+    )
