+++ tests/unit/test_charm.py
@@ -1,0 +1,150 @@
+#!/usr/bin/env python3
+# Copyright 2025 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+from ops import ActiveStatus, pebble, testing
+
+from src.charm import ValkeyK8sCharm
+from src.literals import PEER_RELATION, STATUS_PEERS_RELATION
+from src.statuses import CharmStatuses
+
+from .helpers import status_is
+
+CHARM_USER = "valkey"
+CONTAINER = "valkey"
+SERVICE_VALKEY = "valkey"
+SERVICE_METRIC_EXPORTER = "metric_exporter"
+
+
+def test_pebble_ready_leader_unit(cloud_spec):
+    ctx = testing.Context(ValkeyK8sCharm, app_trusted=True)
+    relation = testing.PeerRelation(id=1, endpoint=PEER_RELATION)
+    status_peer_relation = testing.PeerRelation(id=2, endpoint=STATUS_PEERS_RELATION)
+
+    # happy path
+    container = testing.Container(name=CONTAINER, can_connect=True)
+    state_in = testing.State(
+        model=testing.Model(name="my-vm-model", type="lxd", cloud_spec=cloud_spec),
+        leader=True,
+        relations={relation, status_peer_relation},
+        containers={container},
+    )
+
+    expected_plan = {
+        "services": {
+            SERVICE_VALKEY: {
+                "override": "replace",
+                "summary": "Valkey service",
+                "command": "valkey-server /var/lib/valkey/valkey.conf",
+                "user": CHARM_USER,
+                "group": CHARM_USER,
+                "startup": "enabled",
+            },
+            SERVICE_METRIC_EXPORTER: {
+                "override": "replace",
+                "summary": "Valkey metric exporter",
+                "command": "bin/redis_exporter",
+                "user": CHARM_USER,
+                "group": CHARM_USER,
+                "startup": "enabled",
+            },
+        }
+    }
+
+    state_out = ctx.run(ctx.on.pebble_ready(container), state_in)
+    assert state_out.get_container(container.name).plan == expected_plan
+    assert (
+        state_out.get_container(container.name).service_statuses[SERVICE_VALKEY]
+        == pebble.ServiceStatus.ACTIVE
+    )
+    assert (
+        state_out.get_container(container.name).service_statuses[SERVICE_METRIC_EXPORTER]
+        == pebble.ServiceStatus.ACTIVE
+    )
+    assert state_out.unit_status == ActiveStatus()
+    assert status_is(state_out, CharmStatuses.SCALING_NOT_IMPLEMENTED.value, is_app=True)
+
+    # container not ready
+    container = testing.Container(name=CONTAINER, can_connect=False)
+    state_in = testing.State(
+        model=testing.Model(name="my-vm-model", type="lxd", cloud_spec=cloud_spec),
+        leader=True,
+        relations={relation, status_peer_relation},
+        containers={container},
+    )
+
+    state_out = ctx.run(ctx.on.pebble_ready(container), state_in)
+    assert status_is(state_out, CharmStatuses.SERVICE_NOT_STARTED.value)
+    assert status_is(state_out, CharmStatuses.SERVICE_NOT_STARTED.value, is_app=True)
+
+
+def test_pebble_ready_non_leader_unit(cloud_spec):
+    ctx = testing.Context(ValkeyK8sCharm, app_trusted=True)
+    relation = testing.PeerRelation(id=1, endpoint=PEER_RELATION)
+    status_peer_relation = testing.PeerRelation(id=2, endpoint=STATUS_PEERS_RELATION)
+
+    # happy path
+    container = testing.Container(name=CONTAINER, can_connect=True)
+    state_in = testing.State(
+        model=testing.Model(name="my-vm-model", type="lxd", cloud_spec=cloud_spec),
+        leader=False,
+        relations={relation, status_peer_relation},
+        containers={container},
+    )
+
+    state_out = ctx.run(ctx.on.pebble_ready(container), state_in)
+    assert not state_out.get_container(container.name).service_statuses.get(SERVICE_VALKEY)
+    assert not state_out.get_container(container.name).service_statuses.get(
+        SERVICE_METRIC_EXPORTER
+    )
+    assert status_is(state_out, CharmStatuses.SCALING_NOT_IMPLEMENTED.value)
+
+    # container not ready
+    container = testing.Container(name=CONTAINER, can_connect=False)
+    state_in = testing.State(
+        model=testing.Model(name="my-vm-model", type="lxd", cloud_spec=cloud_spec),
+        leader=True,
+        relations={relation, status_peer_relation},
+        containers={container},
+    )
+
+    state_out = ctx.run(ctx.on.pebble_ready(container), state_in)
+    assert status_is(state_out, CharmStatuses.SERVICE_NOT_STARTED.value)
+
+
+def test_update_status_leader_unit(cloud_spec):
+    ctx = testing.Context(ValkeyK8sCharm, app_trusted=True)
+    relation = testing.PeerRelation(
+        id=1,
+        endpoint=PEER_RELATION,
+        local_unit_data={"started": "True"},
+    )
+    status_peer_relation = testing.PeerRelation(id=2, endpoint=STATUS_PEERS_RELATION)
+
+    # happy path
+    container = testing.Container(name=CONTAINER, can_connect=True)
+    state_in = testing.State(
+        model=testing.Model(name="my-vm-model", type="lxd", cloud_spec=cloud_spec),
+        leader=True,
+        relations={relation, status_peer_relation},
+        containers={container},
+    )
+
+    state_out = ctx.run(ctx.on.update_status(), state_in)
+    assert state_out.unit_status == ActiveStatus()
+
+
+def test_update_status_non_leader_unit(cloud_spec):
+    ctx = testing.Context(ValkeyK8sCharm, app_trusted=True)
+    relation = testing.PeerRelation(id=1, endpoint=PEER_RELATION)
+    status_peer_relation = testing.PeerRelation(id=2, endpoint=STATUS_PEERS_RELATION)
+
+    container = testing.Container(name=CONTAINER, can_connect=True)
+    state_in = testing.State(
+        model=testing.Model(name="my-vm-model", type="lxd", cloud_spec=cloud_spec),
+        leader=False,
+        relations={relation, status_peer_relation},
+        containers={container},
+    )
+    state_out = ctx.run(ctx.on.update_status(), state_in)
+    assert status_is(state_out, CharmStatuses.SCALING_NOT_IMPLEMENTED.value)
