+++ tox.ini
@@ -1,0 +1,72 @@
+# Copyright 2025 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+[tox]
+no_package = True
+skip_missing_interpreters = True
+env_list = format, lint, static, unit
+min_version = 4.0.0
+
+[vars]
+src_path = {tox_root}/src
+tests_path = {tox_root}/tests
+;lib_path = {tox_root}/lib/charms/operator_name_with_underscores
+all_path = {[vars]src_path} {[vars]tests_path}
+
+[testenv]
+set_env =
+    PYTHONPATH = {[vars]src_path}:{tox_root}/lib
+    PY_COLORS = 1
+allowlist_externals =
+    poetry
+
+[testenv:format]
+description = Apply coding style standards to code
+commands_pre =
+    poetry install --only format
+commands =
+    poetry lock
+    poetry run ruff check --fix {[vars]all_path}
+    poetry run ruff format {[vars]all_path}
+
+[testenv:lint]
+description = Check code against coding style standards
+allowlist_externals =
+    {[testenv]allowlist_externals}
+    find
+commands_pre =
+    poetry install --only lint
+commands =
+    poetry check --lock
+    # uncomment the following line if this charm owns a lib
+    # codespell {[vars]lib_path}
+    poetry run codespell {[vars]all_path} --skip "valkey.conf"
+    poetry run ruff check {[vars]all_path}
+    poetry run ruff format --check --diff {[vars]all_path}
+    find {[vars]all_path} -type f \( -name "*.sh" -o -name "*.bash" \) -exec poetry run shellcheck --color=always \{\} +
+
+[testenv:unit]
+description = Run unit tests
+set_env =
+    {[testenv]set_env}
+commands_pre =
+    poetry install --only main,charm-libs,unit
+commands =
+    poetry run coverage run --source={[vars]src_path} --omit=*/lib/charms/* \
+        -m pytest -v --tb native -s {posargs} {[vars]tests_path}/unit
+    poetry run coverage report
+    poetry run coverage xml
+
+[testenv:integration]
+description = Run integration tests
+pass_env =
+    CI
+allowlist_externals =
+    {[testenv]allowlist_externals}
+    sh
+commands_pre =
+    poetry install --only integration
+commands =
+    # on CI, concierge will setup the model `testing` - locally we need to do it ourselves
+    sh -c "if [ -z "$CI" ]; then juju add-model testing; fi;"
+    poetry run pytest -v --tb native --log-cli-level=INFO -s --ignore={[vars]tests_path}/unit/ {posargs}
