--- valkey-cluster-operator/machines/.gitignore
@@ -1,11 +1,0 @@
-venv/
-build/
-*.charm
-.tox/
-.coverage
-__pycache__/
-*.py[cod]
-.idea
-.vscode/
-coverage.xml
-.env
