--- valkey-operator/kubernetes/pyproject.toml
@@ -1,93 +1,0 @@
-# Copyright 2025 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-[tool.poetry]
-package-mode = false
-requires-poetry = ">=2.0.0"
-
-[tool.poetry.dependencies]
-python = "^3.12" # switch to 3.14 once charm base is 26.04
-ops = "^3.5.0"
-charmlibs-pathops = "^1.2.0"
-data-platform-helpers = ">=0.1.7"
-common = {path = "../../common", develop = true}
-
-[tool.poetry.requires-plugins]
-poetry-plugin-export = ">=1.8"
-
-[tool.poetry.group.charm-libs.dependencies]
-# data_platform_libs/v0/data_interfaces.py
-ops = ">=2.0.0"
-
-[tool.poetry.group.format]
-optional = true
-
-[tool.poetry.group.format.dependencies]
-ruff = "*"
-
-[tool.poetry.group.lint]
-optional = true
-
-[tool.poetry.group.lint.dependencies]
-ruff = "*"
-codespell = "*"
-shellcheck-py = "*"
-
-[tool.poetry.group.unit.dependencies]
-pytest = "*"
-pytest-asyncio = "*"
-pytest-mock = "*"
-coverage = { extras = ["toml"], version = "*" }
-parameterized = "*"
-ops-scenario = "*"
-
-[tool.poetry.group.integration.dependencies]
-pytest = "*"
-pytest-asyncio = "*"
-allure-pytest = "*"
-allure-pytest-default-results = "^0.1.2"
-data-platform-helpers = ">=0.1.7"
-jubilant = "^1.6.0"
-
-[tool.coverage.run]
-branch = true
-
-[tool.coverage.report]
-show_missing = true
-exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:"]
-
-[tool.pytest.ini_options]
-minversion = "6.0"
-log_cli_level = "INFO"
-asyncio_mode = "auto"
-asyncio_default_fixture_loop_scope = "function"
-
-# Linting tools configuration
-[tool.ruff]
-line-length = 99
-lint.select = ["E", "W", "F", "C", "N", "D", "I001"]
-lint.extend-ignore = [
-    "D203",
-    "D204",
-    "D213",
-    "D215",
-    "D400",
-    "D404",
-    "D406",
-    "D407",
-    "D408",
-    "D409",
-    "D413",
-]
-lint.ignore = ["E501", "D107"]
-extend-exclude = ["__pycache__", "*.egg_info", "../../common/common/lib/charms/**"]
-lint.per-file-ignores = { "tests/*" = ["D100", "D101", "D102", "D103", "D104"] }
-
-[tool.ruff.lint.mccabe]
-max-complexity = 10
-
-[tool.codespell]
-skip = "build,lib,venv,icon.svg,.tox,.git,.mypy_cache,.ruff_cache,.coverage"
-
-[tool.pyright]
-include = ["src/**/*.py", "lib/**/*.py"]
