--- valkey-operator/kubernetes/tests/spread/test_charm.py/task.yaml
@@ -1,9 +1,0 @@
-summary: test_charm.py
-environment:
-  TEST_MODULE: test_charm.py
-systems:
-  - self-hosted-linux-amd64-noble-medium
-execute: |
-  tox run -e integration -- "tests/integration/$TEST_MODULE" --alluredir="$SPREAD_TASK/allure-results"
-artifacts:
-  - allure-results
